Decompose conversation HTML enhancements into focused hooks
Agent-Logs-Url: https://github.com/jamesmoore/chatgpt-archive-tools/sessions/d9666809-54ef-4f58-be9f-b9dfeea81747
