+++ front-react/src/hooks/use-conversation-html-highlight.ts
@@ -1,0 +1,12 @@
+import { useCallback } from "react";
+import hljs from "highlight.js";
+
+export function useConversationHtmlHighlight() {
+    const highlightCodeBlocks = useCallback((container: HTMLElement) => {
+        container.querySelectorAll("pre code").forEach((block) => {
+            hljs.highlightElement(block as HTMLElement);
+        });
+    }, []);
+
+    return { highlightCodeBlocks };
+}
